Refactored arrayvalidator to use shorthand bind notation for values in constructor. Also moved tests to their own file.
Array validation can now be shortened by writing,
```
Duns.array([100,200]).validate(); //Returns true
Duns.array([100,200]).length(2).validate(); //Returns true
```
